// src/shared/api/subsonic/subsonic-schema.d.ts
export interface components {
    schemas: {
// ...
        ArtistID3: {
            /** @description The id of the artist */
            id: string;
            /** @description The artist name. */
            name: string;
            /** @description A covertArt id. */
            coverArt?: string;
            /** @description An url to an external image source. */
            artistImageUrl?: string;
            /** @description Artist album count. */
            albumCount?: number;
            /**
             * Format: date-time
             * @description Date the artist was starred. [ISO 8601]
             */
            starred?: string;
            /** @description The artist MusicBrainzID. */
            musicBrainzId?: string;
            /** @description The artist sort name. */
            sortName?: string;
            /** @description The list of all roles this artist has in the library. */
            roles?: string[];
        };
// ...
        Playlist: {
            /** @description Id of the playlist */
            id: string;
            /** @description Name of the playlist */
            name: string;
            /** @description A comment */
            comment?: string;
            /** @description Owner of the playlist */
            owner?: string;
            /** @description Is the playlist public */
            public?: boolean;
            /** @description number of songs */
            songCount: number;
            /** @description Playlist duration in seconds */
            duration: number;
            /**
             * Format: date-time
             * @description Creation date [ISO 8601]
             */
            created: string;
            /**
             * Format: date-time
             * @description Last changed date [ISO 8601]
             */
            changed: string;
            /** @description A cover Art Id */
            coverArt?: string;
            /** @description A list of allowed usernames */
            allowedUser?: string[];
        };
// ...
        AlbumID3: {
            /** @description The id of the album */
            id: string;
            /** @description The album name. */
            name: string;
            /** @description The album version name (Remastered, Anniversary Box Set, …). */
            version?: string;
            /** @description Artist name. */
            artist?: string;
            /** @description The id of the artist */
            artistId?: string;
            /** @description A covertArt id. */
            coverArt?: string;
            /** @description Number of songs */
            songCount: number;
            /** @description Total duration of the album in seconds */
            duration: number;
            /** @description Number of play of the album */
            playCount?: number;
            /**
             * Format: date-time
             * @description Date the album was added. [ISO 8601]
             */
            created: string;
            /**
             * Format: date-time
             * @description Date the album was added. [ISO 8601]
             */
            starred?: string;
            /** @description The album year */
            year?: number;
            /** @description The album genre */
            genre?: string;
            /**
             * Format: date-time
             * @description Date the album was last played. [ISO 8601]
             */
            played?: string;
            /** @description The user rating of the album. [1-5] */
            userRating?: number;
            /** @description The labels producing the album. */
            recordLabels?: components["schemas"]["RecordLabel"][];
            /** @description The album MusicBrainzID. */
            musicBrainzId?: string;
            /** @description The list of all genres of the album. */
            genres?: components["schemas"]["ItemGenre"][];
            /** @description The list of all album artists of the album. */
            artists?: components["schemas"]["ArtistID3"][];
            /** @description The single value display artist. */
            displayArtist?: string;
            /** @description The types of this album release. (Album, Compilation, EP, Remix, …). */
            releaseTypes?: string[];
            /** @description The list of all moods of the album. */
            moods?: string[];
            /** @description The album sort name. */
            sortName?: string;
            /** @description Date the album was originally released. */
            originalReleaseDate?: components["schemas"]["ItemDate"];
            /** @description Date the specific edition of the album was released. Note: for files using ID3 tags, releaseDate should generally be read from the TDRL tag. Servers that use a different source for this field should document the behavior. */
            releaseDate?: components["schemas"]["ItemDate"];
            /** @description True if the album is a compilation. */
            isCompilation?: boolean;
            explicitStatus?: components["schemas"]["ExplicitStatus"];
            /** @description The list of all disc titles of the album. */
            discTitles?: components["schemas"]["DiscTitle"][];
            /** @description The list of songs */
            song?: components["schemas"]["Child"][];
        };
        AlbumID3WithSongs: components["schemas"]["AlbumID3"] & {
            /** @description The list of songs */
            song: components["schemas"]["Child"][];
        };
// ...
        ArtistInfo2: {
            /** @description Artist biography */
            biography?: string;
            /** @description Artist musicBrainzId */
            musicBrainzId?: string;
            /** @description Artist lastFmUrl */
            lastFmUrl?: string;
            /** @description Artist smallImageUrl */
            smallImageUrl?: string;
            /** @description Artist mediumImageUrl */
            mediumImageUrl?: string;
            /** @description Artist largeImageUrl */
            largeImageUrl?: string;
            /** @description Similar artists */
            similarArtist?: components["schemas"]["ArtistID3"][];
        };
// ...
        Genre: {
            /** @description Genre name */
            value: string;
            /** @description Genre song count */
            songCount: number;
            /** @description Genre album count */
            albumCount: number;
        };
// ...
    };
// ...
}